fix: keyboard prevent typing

+++ src/modules/editor/navigation/sidebar/sidebar.component.ts
@@ -8,6 +8,7 @@ import { SaveConfirmComponent } from '../dialogs/save-confirm.component';
 import { MatBottomSheet } from '@angular/material/bottom-sheet';
 import { OptionsService } from 'src/modules/core/services/options.service';
 import { AlgorithmsService } from 'src/modules/core/services/algorithms.service';
+import { InputService } from 'src/modules/core/services/input.service';
 
 @Component({
     selector: 'app-sidebar',
@@ -24,45 +25,50 @@ export class SidebarComponent implements OnInit {
         public dialog: MatDialog,
         public bottomSheet: MatBottomSheet,
         public options: OptionsService,
-        public algorithmsService: AlgorithmsService) { }
+        public algorithmsService: AlgorithmsService,
+        public inputService: InputService) { }
     
     @HostListener('window:keydown', ['$event'])
     onKeyDown(event: KeyboardEvent) {
+        if(!this.inputService.isTyping) {
+            if (event.key === "p") {
+                event.preventDefault();
+                this.openFilesMenu();
+            }
 
-        if (event.key === "p") {
-            this.openFilesMenu();
-        }
+            if (event.key === "w") {
+                this.graphService.addNode();
+            }
 
-        if (event.key === "w") {
-            this.graphService.addNode();
-        }
+            if (event.key === "n") {
+                this.openSaveGraphConfirm();
+            }
 
-        if (event.key === "n") {
-            this.openSaveGraphConfirm();
-        }
+            if(event.key === "s") {
+                this.dataService.saveAndCreateNew(this.el, false);
+            }
 
-        if(event.key === "s") {
-            this.dataService.saveAndCreateNew(this.el, false);
-        }
+            if (event.key === "e") {
+                this.graphService.addEdge();
+            }
 
-        if (event.key === "e") {
-            this.graphService.addEdge();
-        }
+            if(event.key === "r") {
+                this.graphService.editEdge();
+            }
 
-        if(event.key === "r") {
-            this.graphService.editEdge();
-        }
+            if(event.shiftKey) {
+                event.preventDefault();
+                this.matrixToggle() ? this.sidebarToggle() : this.matrixToggle();
+            }
 
-        if(event.shiftKey) {
-            this.matrixToggle() ? this.sidebarToggle() : this.matrixToggle();
-        }
+            if(event.altKey) {
+                event.preventDefault();
+                this.algToggle() ? this.sidebarToggle() : this.algToggle();
+            }
 
-        if(event.ctrlKey) {
-            this.algToggle() ? this.sidebarToggle() : this.algToggle();
-        }
-
-        if (event.key === "Delete") {
-            this.graphService.deleteSelected();
+            if (event.key === "Delete") {
+                this.graphService.deleteSelected();
+            }
         }
     }
